+++ src/storage.jl
@@ -1,3 +1,5 @@
+using JLD2
+
 abstract type AbstractInternalStorage end
 
 const _HostTrmat = Vector{Matrix{Float64}}